Create Add Todo feature
Create Add Todo component for adding Todos to store.
When creating a new Todo: id, label and created fields should be added.
Error handling is also in place when trying to add a Todo but it's empty.
When adding a Todo item, add action gets dispatched with the Todo item.